requirements.txt
asgiref==3.4.1
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
uritemplate==3.0.1
urllib3==1.26.6
